Adds underline under menu items
Adding pseudo element(::before), allows us to draw small things.
Like that underline for example. We draw it completely, but make it invisible.
And when user hovers(puts mouse on item) it becames visible.
Careful when using absolute elements!
If previous element isn't set to relative, absolute element will fly of the screen...